Levels Filter (#97)
* Add LevelFilter and reorganize filters

* Add Pillboxes under search bar

* Set up backend request

* Remove console.log in Aggregate

* Use a boolean array instead of a bitfield

* Fix lint error

* Fix crash when levels filter is empty

* Fix levels pillboxes sometimes not appearing

* Expand 1xx-2xx level filter to include 0xx
