+++ frontend/src/app/filters.ts
@@ -17,6 +17,10 @@ export interface FiltersState {
     active: boolean;
     sessions: Session[];
   };
+  levels: {
+    active: boolean;
+    selected: boolean[]; // selected[i] <=> show i00 level
+  };
 }
 
 const initialState: FiltersState = {
@@ -33,6 +37,21 @@ const initialState: FiltersState = {
   semesters: {
     active: false,
     sessions: [],
+  },
+  levels: {
+    active: false,
+    selected: [
+      false, // 000
+      false, // 100
+      false, // 200
+      false, // 300
+      false, // 400
+      false, // 500
+      false, // 600
+      false, // 700
+      false, // 800
+      false, // 900
+    ],
   },
 };
 
@@ -75,6 +94,17 @@ export const filtersSlice = createSlice({
     updateUnitsRange: (state, action: PayloadAction<[number, number]>) => {
       state.units.min = action.payload[0];
       state.units.max = action.payload[1];
+    },
+    updateLevelsActive: (state, action: PayloadAction<boolean>) => {
+      state.levels.active = action.payload;
+    },
+    updateLevelsSelection: (state, action: PayloadAction<boolean[]>) => {
+      state.levels.selected = action.payload;
+    },
+    deleteLevel: (state, action: PayloadAction<number[]>) => {
+      for (const index of action.payload) {
+        state.levels.selected[index] = false;
+      }
     },
   },
 });
