+++ frontend/src/components/Aggregate.tsx
@@ -47,7 +47,6 @@ const NumericInput = (props: NumericInputProps) => {
   // this is only called when a key that creates a character is pressed (for
   // example symbols and alphabets; NOT backspace, arrow keys, Alt/Ctrl etc)
   const handleKeyPress: React.KeyboardEventHandler<HTMLInputElement> = (e) => {
-    console.log("test");
     if (!numericRegex.test(e.key)) e.preventDefault();
   };
