+++ frontend/src/components/SearchBar.tsx
@@ -10,6 +10,7 @@ import {
 } from "../app/utils";
 import { cacheSlice } from "../app/cache";
 import { filtersSlice } from "../app/filters";
+import { getPillboxes } from "./filters/LevelFilter";
 
 const AppliedFiltersPill = ({
   className,
@@ -85,6 +86,22 @@ const AppliedFilters = () => {
           key={`session-${sessionToShortString(session)}`}
         >
           {sessionToString(session)}
+        </AppliedFiltersPill>
+      );
+    });
+  }
+
+  if (filter.levels.active) {
+    getPillboxes(filter.levels.selected).forEach(({ levels, content }) => {
+      badges.push(
+        <AppliedFiltersPill
+          className="text-red-800 bg-red-50"
+          onDelete={() => {
+            dispatch(filtersSlice.actions.deleteLevel(levels));
+          }}
+          key={`session-${levels.toString()}`}
+        >
+          {content}
         </AppliedFiltersPill>
       );
     });
